Add .profile
It turns out ~/.bash_profile is not executed by the X login shell (it seems
to run dash instead of bash). Let ~/.bash_profile simply run ~/.bashrc and
do the same for ~/.profile.

So essentially there are now two "entry points" - one for systems that use
~/.prorfile, other for systems that use ~/.bash_profile (if one of the is
found the other one is usually not run).